move main config parsing into envutils
