+++ packages/indexer/src/main.ts
@@ -1,86 +1,9 @@
-import assert from "assert";
 import * as services from "./services";
 import winston from "winston";
 import Redis from "ioredis";
 import * as across from "@across-protocol/sdk";
 import { connectToDatabase } from "./database/database.provider";
-import { DatabaseConfig } from "@repo/indexer-database";
 import * as utils from "./utils";
-
-type Config = {
-  redisConfig: utils.RedisConfig;
-  postgresConfig: DatabaseConfig;
-  spokeConfigs: Omit<
-    services.spokePoolIndexer.Config,
-    "logger" | "redis" | "postgres"
-  >[];
-  hubConfig: Omit<
-    services.hubPoolIndexer.Config,
-    "logger" | "redis" | "postgres"
-  >;
-};
-
-export function envToConfig(env: utils.Env): Config {
-  assert(env.HUBPOOL_CHAIN, "Requires HUBPOOL_CHAIN");
-  const redisConfig = utils.parseRedisConfig(env);
-  const postgresConfig = utils.parsePostgresConfig(env);
-  const allProviderConfigs = utils.parseProviderConfigs(env);
-  const retryProviderConfig = utils.parseRetryProviderConfig(env);
-  const hubPoolChain = utils.parseNumber(env.HUBPOOL_CHAIN);
-  const spokePoolChainsEnabled = utils
-    .parseArray(env.SPOKEPOOL_CHAINS_ENABLED)
-    .map(utils.parseNumber);
-  const providerConfigs = allProviderConfigs.filter(
-    (provider) => provider[1] === hubPoolChain,
-  );
-  assert(
-    allProviderConfigs.length > 0,
-    `Requires at least one RPC_PROVIDER_URLS_CHAINID`,
-  );
-
-  const hubConfig = {
-    retryProviderConfig: {
-      ...retryProviderConfig,
-      chainId: hubPoolChain,
-      providerConfigs,
-    },
-    hubConfig: {
-      chainId: hubPoolChain,
-      maxBlockLookBack: 10000,
-    },
-    redisKeyPrefix: `hubPoolIndexer:${hubPoolChain}`,
-  };
-
-  const spokeConfigs = spokePoolChainsEnabled.map((chainId) => {
-    const providerConfigs = allProviderConfigs.filter(
-      (provider) => provider[1] == chainId,
-    );
-    assert(
-      providerConfigs.length > 0,
-      `SPOKEPOOL_CHAINS_ENABLED=${chainId} but did not find any corresponding RPC_PROVIDER_URLS_${chainId}`,
-    );
-    return {
-      retryProviderConfig: {
-        ...retryProviderConfig,
-        chainId,
-        providerConfigs,
-      },
-      spokeConfig: {
-        chainId,
-        maxBlockLookBack: 10000,
-      },
-      hubConfig: hubConfig.hubConfig,
-      redisKeyPrefix: `spokePoolIndexer:${chainId}`,
-    };
-  });
-
-  return {
-    redisConfig,
-    postgresConfig,
-    hubConfig,
-    spokeConfigs,
-  };
-}
 
 async function initializeRedis(
   config: utils.RedisConfig,
@@ -103,7 +26,7 @@ async function initializeRedis(
   });
 }
 
-export async function Main(config: Config, logger: winston.Logger) {
+export async function Main(config: utils.Config, logger: winston.Logger) {
   const { redisConfig, postgresConfig, hubConfig, spokeConfigs } = config;
 
   const redis = await initializeRedis(redisConfig, logger);
