+++ packages/indexer/src/utils/envUtils.ts
@@ -2,6 +2,20 @@ import assert from "assert";
 import { DatabaseConfig } from "@repo/indexer-database";
 import * as s from "superstruct";
 import { RetryProviderConfig } from "./contractUtils";
+import * as services from "../services";
+
+export type Config = {
+  redisConfig: RedisConfig;
+  postgresConfig: DatabaseConfig;
+  spokeConfigs: Omit<
+    services.spokePoolIndexer.Config,
+    "logger" | "redis" | "postgres"
+  >[];
+  hubConfig: Omit<
+    services.hubPoolIndexer.Config,
+    "logger" | "redis" | "postgres"
+  >;
+};
 export type RedisConfig = {
   host: string;
   port: number;
@@ -10,7 +24,7 @@ export type ProviderConfig = [providerUrl: string, chainId: number];
 
 export type Env = Record<string, string | undefined>;
 
-export function parseRedisConfig(env: Env): RedisConfig {
+function parseRedisConfig(env: Env): RedisConfig {
   assert(env.REDIS_HOST, "requires REDIS_HOST");
   assert(env.REDIS_PORT, "requires REDIS_PORT");
   const port = parseNumber(env.REDIS_PORT);
@@ -20,7 +34,7 @@ export function parseRedisConfig(env: Env): RedisConfig {
   };
 }
 
-export function parseArray(value: string | undefined): string[] {
+function parseArray(value: string | undefined): string[] {
   if (value === undefined) return [];
   return value
     .split(",")
@@ -29,14 +43,14 @@ export function parseArray(value: string | undefined): string[] {
 }
 
 // superstruct coersion to turn string into an int and validate
-export const stringToInt = s.coerce(s.number(), s.string(), (value) =>
+const stringToInt = s.coerce(s.number(), s.string(), (value) =>
   parseInt(value),
 );
-export function parseNumber(value: string): number {
+function parseNumber(value: string): number {
   return s.create(value, stringToInt);
 }
 
-export function parsePostgresConfig(
+function parsePostgresConfig(
   env: Record<string, string | undefined>,
 ): DatabaseConfig {
   assert(env.DATABASE_HOST, "requires DATABASE_HOST");
@@ -52,7 +66,7 @@ export function parsePostgresConfig(
     dbName: env.DATABASE_NAME,
   };
 }
-export function parseProviderConfigs(env: Env): ProviderConfig[] {
+function parseProviderConfigs(env: Env): ProviderConfig[] {
   const results: ProviderConfig[] = [];
   for (const [key, value] of Object.entries(process.env)) {
     const match = key.match(/^RPC_PROVIDER_URLS_(\d+)$/);
@@ -69,7 +83,7 @@ export function parseProviderConfigs(env: Env): ProviderConfig[] {
   return results;
 }
 
-export function parseRetryProviderConfig(
+function parseRetryProviderConfig(
   env: Record<string, string | undefined>,
 ): Omit<RetryProviderConfig, "providerConfigs" | "chainId"> {
   assert(env.PROVIDER_CACHE_NAMESPACE, "requires PROVIDER_CACHE_NAMESPACE");
@@ -98,5 +112,67 @@ export function parseRetryProviderConfig(
     nodeQuorumThreshold: s.create(env.NODE_QUORUM_THRESHOLD, stringToInt),
     retries: s.create(env.RETRIES, stringToInt),
     delay: s.create(env.DELAY, stringToInt),
+  };
+}
+
+export function envToConfig(env: Env): Config {
+  assert(env.HUBPOOL_CHAIN, "Requires HUBPOOL_CHAIN");
+  const redisConfig = parseRedisConfig(env);
+  const postgresConfig = parsePostgresConfig(env);
+  const allProviderConfigs = parseProviderConfigs(env);
+  const retryProviderConfig = parseRetryProviderConfig(env);
+  const hubPoolChain = parseNumber(env.HUBPOOL_CHAIN);
+  const spokePoolChainsEnabled = parseArray(env.SPOKEPOOL_CHAINS_ENABLED).map(
+    parseNumber,
+  );
+  const providerConfigs = allProviderConfigs.filter(
+    (provider) => provider[1] === hubPoolChain,
+  );
+  assert(
+    allProviderConfigs.length > 0,
+    `Requires at least one RPC_PROVIDER_URLS_CHAINID`,
+  );
+
+  const hubConfig = {
+    retryProviderConfig: {
+      ...retryProviderConfig,
+      chainId: hubPoolChain,
+      providerConfigs,
+    },
+    hubConfig: {
+      chainId: hubPoolChain,
+      maxBlockLookBack: 10000,
+    },
+    redisKeyPrefix: `hubPoolIndexer:${hubPoolChain}`,
+  };
+
+  const spokeConfigs = spokePoolChainsEnabled.map((chainId) => {
+    const providerConfigs = allProviderConfigs.filter(
+      (provider) => provider[1] == chainId,
+    );
+    assert(
+      providerConfigs.length > 0,
+      `SPOKEPOOL_CHAINS_ENABLED=${chainId} but did not find any corresponding RPC_PROVIDER_URLS_${chainId}`,
+    );
+    return {
+      retryProviderConfig: {
+        ...retryProviderConfig,
+        chainId,
+        providerConfigs,
+      },
+      spokeConfig: {
+        chainId,
+        maxBlockLookBack: 10000,
+      },
+      hubConfig: hubConfig.hubConfig,
+      redisKeyPrefix: `spokePoolIndexer:${chainId}`,
+    };
+  });
+
+  return {
+    redisConfig,
+    postgresConfig,
+    hubConfig,
+    spokeConfigs,
   };
 }
